Separate addDays() into new helpers module
